Move encodeOrder and decodeOrder for Wasm compatibility
