+++ core/message_handler.go
@@ -3,8 +3,6 @@
 package core
 
 import (
-	"encoding/json"
-	"fmt"
 	"math/rand"
 
 	"github.com/0xProject/0x-mesh/meshdb"
@@ -13,29 +11,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	log "github.com/sirupsen/logrus"
 )
-
-type orderMessage struct {
-	MessageType string
-	Order       *zeroex.SignedOrder
-}
-
-func encodeOrder(order *zeroex.SignedOrder) ([]byte, error) {
-	return json.Marshal(orderMessage{
-		MessageType: "order",
-		Order:       order,
-	})
-}
-
-func decodeOrder(data []byte) (*zeroex.SignedOrder, error) {
-	var orderMessage orderMessage
-	if err := json.Unmarshal(data, &orderMessage); err != nil {
-		return nil, err
-	}
-	if orderMessage.MessageType != "order" {
-		return nil, fmt.Errorf("unexpected message type: %q", orderMessage.MessageType)
-	}
-	return orderMessage.Order, nil
-}
 
 // Ensure that App implements p2p.MessageHandler.
 var _ p2p.MessageHandler = &App{}
